OBPIH-4921 Fix my stock movements case and add comment to useEffect (code review fix)
