+++ src/js/components/stock-movement/inbound/StockMovementInboundFilters.jsx
@@ -1,141 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React, { useEffect } from 'react';
 
 import PropTypes from 'prop-types';
 import { connect } from 'react-redux';
 
 import { fetchShipmentStatusCodes } from 'actions';
 import FilterForm from 'components/Filter/FilterForm';
-import DateFilter from 'components/form-elements/DateFilter/DateFilter';
-import FilterSelectField from 'components/form-elements/FilterSelectField';
 import { debounceLocationsFetch, debounceUsersFetch } from 'utils/option-utils';
-
-const filterFields = {
-  receiptStatusCode: {
-    type: FilterSelectField,
-    attributes: {
-      multi: true,
-      filterElement: true,
-      placeholder: 'Receipt Status',
-      showLabelTooltip: true,
-      closeMenuOnSelect: false,
-    },
-    getDynamicAttr: ({ shipmentStatuses }) => ({
-      options: shipmentStatuses,
-    }),
-  },
-  origin: {
-    type: FilterSelectField,
-    attributes: {
-      async: true,
-      openOnClick: false,
-      autoload: false,
-      cache: false,
-      valueKey: 'id',
-      labelKey: 'name',
-      options: [],
-      filterOptions: options => options,
-      filterElement: true,
-      placeholder: 'Origin',
-      showLabelTooltip: true,
-    },
-    getDynamicAttr: ({
-      fetchLocations,
-    }) => ({
-      loadOptions: fetchLocations,
-    }),
-  },
-  destination: {
-    type: FilterSelectField,
-    attributes: {
-      valueKey: 'id',
-      filterElement: true,
-      placeholder: 'Destination',
-      showLabelTooltip: true,
-      disabled: true,
-    },
-  },
-  requestedBy: {
-    type: FilterSelectField,
-    attributes: {
-      async: true,
-      openOnClick: false,
-      autoload: false,
-      cache: false,
-      valueKey: 'id',
-      labelKey: 'name',
-      options: [],
-      filterOptions: options => options,
-      filterElement: true,
-      placeholder: 'Requested By',
-      showLabelTooltip: true,
-    },
-    getDynamicAttr: ({
-      fetchUsers,
-    }) => ({
-      loadOptions: fetchUsers,
-    }),
-  },
-  createdBy: {
-    type: FilterSelectField,
-    attributes: {
-      async: true,
-      openOnClick: false,
-      autoload: false,
-      cache: false,
-      valueKey: 'id',
-      labelKey: 'name',
-      options: [],
-      filterOptions: options => options,
-      filterElement: true,
-      placeholder: 'Created By',
-      showLabelTooltip: true,
-    },
-    getDynamicAttr: ({
-      fetchUsers,
-    }) => ({
-      loadOptions: fetchUsers,
-    }),
-  },
-  updatedBy: {
-    type: FilterSelectField,
-    attributes: {
-      async: true,
-      openOnClick: false,
-      autoload: false,
-      cache: false,
-      valueKey: 'id',
-      labelKey: 'name',
-      options: [],
-      filterOptions: options => options,
-      filterElement: true,
-      placeholder: 'Updated By',
-      showLabelTooltip: true,
-    },
-    getDynamicAttr: ({
-      fetchUsers,
-    }) => ({
-      loadOptions: fetchUsers,
-    }),
-  },
-  createdAfter: {
-    type: DateFilter,
-    attributes: {
-      label: 'react.stockMovement.filter.createdAfter.label',
-      defaultMessage: 'Created after',
-      dateFormat: 'MM/DD/YYYY',
-      filterElement: true,
-    },
-  },
-  createdBefore: {
-    type: DateFilter,
-    attributes: {
-      label: 'react.stockMovement.filter.createdBefore.label',
-      defaultMessage: 'Created before',
-      dateFormat: 'MM/DD/YYYY',
-      filterElement: true,
-    },
-  },
-};
 
 const StockMovementInboundFilters = ({
   setFilterParams,
@@ -144,32 +14,17 @@ const StockMovementInboundFilters = ({
   fetchStatuses,
   shipmentStatuses,
   isShipmentStatusesFetched,
-  currentLocation,
+  filterFields,
+  defaultValues,
 }) => {
   const fetchUsers = debounceUsersFetch(debounceTime, minSearchLength);
   const fetchLocations = debounceLocationsFetch(debounceTime, minSearchLength, [], true);
-  const [defaultValues, setDefaultValues] = useState({});
 
 
   useEffect(() => {
     if (!isShipmentStatusesFetched) fetchStatuses();
   }, []);
 
-  useEffect(() => {
-    const initialEmptyValues = Object.keys(filterFields).reduce((acc, key) => {
-      if (!acc[key]) return { ...acc, [key]: '' };
-      return acc;
-    }, {});
-    setDefaultValues({
-      ...initialEmptyValues,
-      destination: {
-        id: currentLocation?.id,
-        value: currentLocation?.id,
-        name: currentLocation?.name,
-        label: currentLocation?.name,
-      },
-    });
-  }, [currentLocation]);
 
   return (
     <div className="d-flex flex-column list-page-filters">
@@ -178,6 +33,7 @@ const StockMovementInboundFilters = ({
         searchFieldPlaceholder="Search by order number of description"
         filterFields={filterFields}
         defaultValues={defaultValues}
+        onClear={form => form.reset({ destination: defaultValues.destination })}
         updateFilterParams={values => setFilterParams({ ...values })}
         hidden={false}
         formProps={{
@@ -195,7 +51,6 @@ const mapStateToProps = state => ({
   minSearchLength: state.session.searchConfig.minSearchLength,
   shipmentStatuses: state.shipmentStatuses.data,
   isShipmentStatusesFetched: state.shipmentStatuses.fetched,
-  currentLocation: state.session.currentLocation,
 });
 
 const mapDispatchToProps = {
@@ -217,8 +72,6 @@ StockMovementInboundFilters.propTypes = {
     variant: PropTypes.string.isRequired,
     label: PropTypes.string.isRequired,
   }).isRequired,
-  currentLocation: PropTypes.shape({
-    id: PropTypes.string.isRequired,
-    name: PropTypes.string.isRequired,
-  }).isRequired,
+  filterFields: PropTypes.shape({}).isRequired,
+  defaultValues: PropTypes.shape({}).isRequired,
 };
